requirements.txt
pydantic>=2.0
google-genai>=1.0
